chore: 타입스크립트로 세팅
